Fix part of the test
